requirements.txt
astroid==3.0.1
attrs==23.1.0
aws-cdk-lib==2.104.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.lambda-layer-kubectl-v24==2.0.242
blinker==1.7.0
cattrs==23.1.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constructs==10.3.0
coverage==7.3.2
dill==0.3.7
exceptiongroup==1.1.3
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsii==1.91.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.18.0
prometheus-flask-exporter==0.23.0
psycopg2-binary==2.9.9
publication==0.0.3
pylint==3.0.2
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
tomlkit==0.12.2
typeguard==2.13.3
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
Flask-Migrate
gunicorn==21.2.0